moved mock config data to adapters folder
