+++ src/Cards/CompositeCards/ADT3DScenePage/ADT3DScenePage.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import useAuthParams from '../../../../.storybook/useAuthParams';
 import MockAdapter from '../../../Adapters/MockAdapter';
 import ADT3DScenePage from './ADT3DScenePage';
-import { mockVConfig } from '../../SceneListCard/Consume/__mockData__/vconfigDecFinal';
+import { mockVConfig } from '../../../Adapters/__mockData__/vconfigDecFinal';
 
 export default {
     title: 'CompositeCards/ADT3DScenePage'
@@ -13,7 +13,7 @@ const cardStyle = {
     width: '100%'
 };
 
-export const ADT3DScenePageCard = (_args, { globals: { theme, locale } }) => {
+export const Mock3DScenePageCard = (_args, { globals: { theme, locale } }) => {
     const authenticationParameters = useAuthParams();
 
     return !authenticationParameters ? (
