+++ src/Cards/SceneListCard/Consume/SceneListCard.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import SceneListCard from './SceneListCard';
 import MockAdapter from '../../../Adapters/MockAdapter';
-import { mockVConfig } from './__mockData__/vconfigDecFinal';
+import { mockVConfig } from '../../../Adapters/__mockData__/vconfigDecFinal';
 
 export default {
     title: 'SceneListCard/Consume'
